feat(clients): watch toggle, history page, pagination, and component refactor
